Struct aws_sdk_mwaa::client::fluent_builders::CreateEnvironment

pub struct CreateEnvironment<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateEnvironment<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateEnvironmentOutput, SdkError<CreateEnvironmentError>> where
R::Policy: SmithyRetryPolicy<CreateEnvironmentInputOperationOutputAlias, CreateEnvironmentOutput, CreateEnvironmentError, CreateEnvironmentInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

pub fn set_name(self, input: Option<String>) -> Self

pub fn execution_role_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

pub fn set_execution_role_arn(self, input: Option<String>) -> Self

pub fn source_bucket_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

pub fn set_source_bucket_arn(self, input: Option<String>) -> Self

pub fn dag_s3_path(self, inp: impl Into<String>) -> Self

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

pub fn set_dag_s3_path(self, input: Option<String>) -> Self

pub fn network_configuration(self, inp: NetworkConfiguration) -> Self

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn plugins_s3_path(self, inp: impl Into<String>) -> Self

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

pub fn set_plugins_s3_path(self, input: Option<String>) -> Self

pub fn plugins_s3_object_version(self, inp: impl Into<String>) -> Self

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

pub fn set_plugins_s3_object_version(self, input: Option<String>) -> Self

pub fn requirements_s3_path(self, inp: impl Into<String>) -> Self

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

pub fn set_requirements_s3_path(self, input: Option<String>) -> Self

pub fn requirements_s3_object_version(self, inp: impl Into<String>) -> Self

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

pub fn set_requirements_s3_object_version(self, input: Option<String>) -> Self

pub fn airflow_configuration_options(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to AirflowConfigurationOptions.

To override the contents of this collection use set_airflow_configuration_options.

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

pub fn set_airflow_configuration_options(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn environment_class(self, inp: impl Into<String>) -> Self

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

pub fn set_environment_class(self, input: Option<String>) -> Self

pub fn max_workers(self, inp: i32) -> Self

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

pub fn set_max_workers(self, input: Option<i32>) -> Self

pub fn kms_key(self, inp: impl Into<String>) -> Self

The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

pub fn set_kms_key(self, input: Option<String>) -> Self

pub fn airflow_version(self, inp: impl Into<String>) -> Self

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

pub fn set_airflow_version(self, input: Option<String>) -> Self

pub fn logging_configuration(self, inp: LoggingConfigurationInput) -> Self

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

pub fn set_logging_configuration(
self,
input: Option<LoggingConfigurationInput>
) -> Self

pub fn weekly_maintenance_window_start(self, inp: impl Into<String>) -> Self

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

pub fn set_weekly_maintenance_window_start(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn webserver_access_mode(self, inp: WebserverAccessMode) -> Self

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

pub fn set_webserver_access_mode(
self,
input: Option<WebserverAccessMode>
) -> Self

pub fn min_workers(self, inp: i32) -> Self

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

pub fn set_min_workers(self, input: Option<i32>) -> Self

pub fn schedulers(self, inp: i32) -> Self

The number of Apache Airflow schedulers to run in your environment.

pub fn set_schedulers(self, input: Option<i32>) -> Self