Module aws_sdk_mwaa::model

Modules

dimension

See Dimension

environment

See Environment

last_update

See LastUpdate

logging_configuration

See LoggingConfiguration

logging_configuration_input

See LoggingConfigurationInput

metric_datum

See MetricDatum

module_logging_configuration

See ModuleLoggingConfiguration

module_logging_configuration_input

See ModuleLoggingConfigurationInput

network_configuration

See NetworkConfiguration

statistic_set

See StatisticSet

update_error

See UpdateError

update_network_configuration_input

See UpdateNetworkConfigurationInput

Structs

Dimension

Internal only API.

Environment

The Amazon Managed Workflows for Apache Airflow (MWAA) environment.

LastUpdate

The status of the last update on the environment, and any errors that were encountered.

LoggingConfiguration

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

LoggingConfigurationInput

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

MetricDatum

Internal only API.

ModuleLoggingConfiguration

Defines the type of logs to send for the Apache Airflow log type (e.g. DagProcessingLogs). Valid values: CloudWatchLogGroupArn, Enabled, LogLevel.

ModuleLoggingConfigurationInput

Defines the type of logs to send for the Apache Airflow log type (e.g. DagProcessingLogs). Valid values: CloudWatchLogGroupArn, Enabled, LogLevel.

NetworkConfiguration

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

StatisticSet

Internal only API.

UpdateError

An object containing the error encountered with the last update: ErrorCode, ErrorMessage.

UpdateNetworkConfigurationInput

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

Enums

EnvironmentStatus

LoggingLevel

Unit

Unit

UpdateStatus

WebserverAccessMode